requirements.txt
absl-py==0.7.1
alabaster==0.7.11
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.1
Babel==2.6.0
backcall==0.1.0
batchgenerators==0.18.2
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cryptography==2.3
cycler==0.10.0
Cython==0.29.6
dask==0.18.1
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
grpcio==1.13.0
h5py==2.9.0
html5lib==1.0.1
idna==2.7
imagesize==1.0.0
isort==4.3.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==2.6.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
linecache2==1.0.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nibabel==2.3.0
nilearn==0.4.2
numpy==1.14.5
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
packaging==17.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.2.1
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.4.6
ptyprocess==0.6.0
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.0.1
PyOpenGL==3.1.0
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
pyzmq==17.1.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
requests==2.22.0
rope==0.14.0
scikit-image==0.14.0
scikit-learn==0.21.3
scipy==1.1.0
SecretStorage==3.0.1
simplegeneric==0.8.1
SimpleITK==1.2.2
sip==4.19.8
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
testpath==0.3.1
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1
traceback2==1.4.0
traitlets==4.3.2
typed-ast==1.1.0
unittest2==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wrapt==1.10.11
